requirements.txt
anyio==4.11.0
certifi==2025.11.12
emoji==2.15.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hypothesis==6.148.1
idna==3.11
jsonschema==4.25.1
markdown-it-py==4.0.0
mdurl==0.1.2
psutil==7.1.3
Pygments==2.19.2
PyYAML==6.0.3
rich==14.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
pytest-asyncio==1.3.0
